чистка кода eslint, разделение игрового экрана снова на три модуля
