+++ js/data.js
@@ -1,67 +1,3 @@
-
-//const levels = [
-//  {
-//    levelType: `one-card`,
-//    levelNumber: 1,
-//    cards: [
-//      {
-//        cardContent: `http://placehold.it/468x458`,
-//        answers: [
-//          { title: `photo`, isCorrect: true},
-//          { title: `paint`, isCorrect: false}
-//        ]
-//      }
-//    ]
-//  },
-//  {
-//    levelType: `two-cards`,
-//    levelNumber: 2,
-//    cards: [
-//      {
-//        cardContent: `http://placehold.it/468x458`,
-//        answers: [
-//          { title: `photo`, isCorrect: true},
-//          { title: `paint`, isCorrect: false}
-//        ]
-//      },
-//      {
-//        cardContent: `http://placehold.it/468x458`,
-//        answers: [
-//          { title: `photo`, isCorrect: true},
-//          { title: `paint`, isCorrect: false}
-//        ]
-//      }
-//    ]
-//  },
-//  {
-//    levelType: `three-cards`,
-//    levelNumber: 3,
-//    cards: [
-//      {
-//        cardContent: `http://placehold.it/468x458`,
-//        answers: [
-//          { title: `photo`, isCorrect: true},
-//          { title: `paint`, isCorrect: false}
-//        ]
-//      },
-//      {
-//        cardContent: `http://placehold.it/468x458`,
-//        answers: [
-//          { title: `photo`, isCorrect: true},
-//          { title: `paint`, isCorrect: false}
-//        ]
-//      },
-//      {
-//        cardContent: `http://placehold.it/468x458`,
-//        answers: [
-//          { title: `photo`, isCorrect: true},
-//          { title: `paint`, isCorrect: false}
-//        ]
-//      }
-//    ]
-//  }
-//];
-
 const levels = [
   {
     levelType: `two-cards`,
@@ -287,54 +223,6 @@ let userState = {
   lives: 3
 };
 
-//let answers = [
-//  {
-//    solved: false,
-//    time: 5000
-//  },
-//  {
-//    solved: true,
-//    time: 22000
-//  },
-//  {
-//    solved: true,
-//    time: 7000
-//  },
-//  {
-//    solved: true,
-//    time: 15000
-//  },
-//  {
-//    solved: true,
-//    time: 12000
-//  },
-//  {
-//    solved: true,
-//    time: 17000
-//  },
-//  {
-//    solved: true,
-//    time: 26000
-//  },
-//  {
-//    solved: true,
-//    time: 12000
-//  },
-//  {
-//    solved: true,
-//    time: 7000
-//  },
-//  {
-//    solved: true,
-//    time: 11000
-//  }
-//];
-
 let answers = [];
-
-//let state = {
-//  questions: [],
-//  answers: []
-//};
 
 export {initialState, levels, answers, userState};
